requirements.txt
streamlit~=1.31.1
torch~=2.3.0
Pillow~=10.1.0
torchvision~=0.17.0